% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_tidy_morse}
\alias{data_tidy_morse}
\title{Morse Code Data}
\format{
A data frame with 36 rows and 36 columns, where
each element is the proportion of times that the row character was confused
with the column character by people unfamiliar with Morse code.
}
\source{
morse.code.csv
}
\usage{
data_tidy_morse
}
\description{
A dissimilarity matrix of rates of confusion between Morse code characters.
}
\keyword{datasets}
