% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_tidy_sa_distance}
\alias{data_tidy_sa_distance}
\title{Distances between South African cities}
\format{
A data frame with 12 rows and 12 columns.
}
\source{
SAdistances.xlsx
}
\usage{
data_tidy_sa_distance
}
\description{
Proximity matrix of distances (in km) between South African cities.
}
\keyword{datasets}
